feat: added a basic view for viewing meal information feat: added a loading spinner in order to provide slow network feedback to the user when loading the desserts list
